requirements.txt
python-dotenv>=1.0.0
dnspython>=2.6.0
email-validator>=2.1.0
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
loguru>=0.7.2
